+++ cors/cors.go
@@ -124,5 +124,5 @@ func (o *Options) HandlePreflight(headers http.Header, requestHeaders http.Heade
 func (o *Options) validateOrigin(requestHeaders http.Header) bool {
 	return len(o.AllowedOrigins) == 0 ||
 		o.AllowedOrigins[0] == "*" ||
-		helper.Contains(o.AllowedOrigins, requestHeaders.Get("Origin"))
+		helper.ContainsStr(o.AllowedOrigins, requestHeaders.Get("Origin"))
 }